requirements.txt
asgiref==3.11.1
Django==6.0.3
pillow==12.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
Pillow>=11.0.0
